Refactor DOM advanced TC1 helpers
